created custom hooks named useErrorModal
